chapter 4 updates
