--- 04_Building Convolutional Neural Networks/01_Extract images from the disk/README.md
@@ -1,1 +1,0 @@
-
