--- 04_Building Convolutional Neural Networks/02_Create image variations for train data/README.md
@@ -1,1 +1,0 @@
-
