--- 04_Building Convolutional Neural Networks/03_Image pre-processing and design of input layers/README.md
@@ -1,1 +1,0 @@
-
