--- 04_Building Convolutional Neural Networks/04_Constructing hidden layers for CNN/README.md
@@ -1,1 +1,0 @@
-
