--- 04_Building Convolutional Neural Networks/05_Constructing output layers for output classification/README.md
@@ -1,1 +1,0 @@
-
