--- 04_Building Convolutional Neural Networks/06_Training images and evaluating CNN output/README.md
@@ -1,1 +1,0 @@
-
