--- 04_Building Convolutional Neural Networks/sourceCode/spring-dl4j/src/main/java/com/springdl4j/springdl4j/service/CookBookService.java
@@ -1,11 +1,0 @@
-package com.springdl4j.springdl4j.service;
-
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-import java.util.List;
-
-
-public interface CookBookService {
-    List<String> generateStringOutput(MultipartFile multipartFile) throws IOException, InterruptedException;
-}
