+++ 04_Building_Convolutional_Neural_Networks/sourceCode/cookbookapp/.gitignore
