+++ 04_Building_Convolutional_Neural_Networks/sourceCode/spring-dl4j/src/main/java/com/springdl4j/springdl4j/service/CookBookService.java
@@ -1,0 +1,11 @@
+package com.springdl4j.springdl4j.service;
+
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+import java.util.List;
+
+
+public interface CookBookService {
+    List<String> generateStringOutput(MultipartFile multipartFile, String modelFilePath) throws IOException, InterruptedException;
+}
