+++ 04_Building_Convolutional_Neural_Networks/sourceCode/spring-dl4j/src/main/resources/application.properties
